Fix a bug when the previous sleep date was in the previous month.
The integration was only querying for the current day from myAir, but myAir would return the whole month.
This doesn't help on the first of the month, because we request for Jan 1, myAir returns data for all of Jan, but this is no data as the last sleep day was 12/31.

Now the script always explicitly asks for the previous 30 days of data. This means if you haven't used your CPAP in 30 days your sensors will be unavailable, which seems an unlikely scenario for anyone who cares enough about their CPAP to be using it with HA.