[2]Add experiment EU support
This change adds experimental EU support.
This is difficult for me to test as I do not have an EU-registered device, but this should work based upon information gathered from the community in the associated issue.
Please report success or failures(along with associated logs) to issue #2 in the repo