Support multiple integration/logins that are in the EU.
The device serial number is used as the unique Id, but we were using 'Unknown' in the EU because the serial number does not appear on the EU dashboard.

Now, the integration uses the username as the serial number which should allow multiple instances of the integration to be added.

Testing:

The unit test for EU is updated, but I don't have an EU account, let alone 2! so I can't properly test